+++ src/main/java/net/shortninja/staffplus/core/domain/staff/alerts/handlers/NameChangeAlertHandler.java
@@ -7,12 +7,17 @@ import net.shortninja.staffplus.core.application.session.OnlineSessionsManager;
 import net.shortninja.staffplus.core.common.permissions.PermissionHandler;
 import net.shortninja.staffplus.core.domain.player.PlayerManager;
 import net.shortninja.staffplus.core.domain.player.settings.PlayerSettingsRepository;
+import net.shortninja.staffplus.core.domain.player.namechanged.bungee.NameChangeBungeeDto;
+import net.shortninja.staffplus.core.domain.player.namechanged.bungee.NameChangedBungeeEvent;
 import net.shortninja.staffplus.core.domain.staff.alerts.config.AlertsConfiguration;
 import net.shortninja.staffplusplus.alerts.AlertType;
 import net.shortninja.staffplusplus.chat.NameChangeEvent;
+import net.shortninja.staffplusplus.session.SppPlayer;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+
+import java.util.Optional;
 
 @IocBean
 @IocListener
@@ -38,8 +43,29 @@ public class NameChangeAlertHandler extends AlertsHandler implements Listener {
         if (permission.has(nameChangeEvent.getPlayer(), alertsConfiguration.permissionNameChangeBypass)) {
             return;
         }
+        notifyPlayers(nameChangeEvent.getOldName(), nameChangeEvent.getNewName(), nameChangeEvent.getServerName());
+    }
+
+    @EventHandler
+    public void handle(NameChangedBungeeEvent nameChangeEvent) {
+        if (!alertsConfiguration.alertsNameNotify) {
+            return;
+        }
+        NameChangeBungeeDto nameChangeBungeeDto = nameChangeEvent.getNameChangeBungeeDto();
+
+        Optional<SppPlayer> sppPlayer = playerManager.getOnOrOfflinePlayer(nameChangeBungeeDto.getPlayerUuid());
+        if (sppPlayer.isPresent() && permission.has(sppPlayer.get().getOfflinePlayer(), alertsConfiguration.permissionNameChangeBypass)) {
+            return;
+        }
+        notifyPlayers(nameChangeBungeeDto.getOldName(), nameChangeBungeeDto.getNewName(), nameChangeBungeeDto.getServerName());
+    }
+
+    private void notifyPlayers(String oldName, String newName, String serverName) {
         for (Player player : getPlayersToNotify()) {
-            messages.send(player, messages.alertsName.replace("%old%", nameChangeEvent.getOldName()).replace("%new%", nameChangeEvent.getNewName()), messages.prefixGeneral, getPermission());
+            messages.send(player, messages.alertsName
+                .replace("%old%", oldName)
+                .replace("%server%", serverName)
+                .replace("%new%", newName), messages.prefixGeneral, getPermission());
         }
     }