+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - Add bans-migrate command (migrate default bukkit ipbans to staff++)
   - Add configured commands on chat phrase detection
   - The report command can now be used from console.
+  - Chat phrases detection can now have actions configured
+  - Added command detection.
 
 ### Bugs
   - Fix potion effects not wearing off if a frozen player logs out.